♻️  Refactor: Rename queries module to queries-scheduling-options and update imports
